Address review: locked id allocation, double arithmetic, exact traceability, detail probe
SettlementRepository.save allocates max(settlement_id)+1 and inserts under one lock (SETTLE-R12); SettlementRepositoryTest runs 32 concurrent saves and expects 32 distinct ids.

LegacyCoercions parses with Double.parseDouble exactly as ClaimsActionSupport.decimal does and the arithmetic is double end to end (SETTLE-R03 v2, SETTLE-R09 v2, QUIRK-08); SPEC-SETTLE-001 bumped to 0.3 with history rows. LegacyErrorHandler no longer adds errors.system, matching the legacy error screen (SETTLE-R06, QUIRK-05).

tools/traceability.py matches rule ids exactly, so SETTLE-R06 and SETTLE-R06 v2 get their own rows and commits. parity/replay.py and tools/capture/capture.py probe /settlement/detail and read detailAmount for the save scenario (SETTLE-R13); transcript re-captured.

CI runs service tests and parity and checks parity/report.json, parity/report.md and docs/TRACEABILITY.md with git diff --exit-code.

ADR-001 decision 8 records why authentication and input length stay as they are; OQ-04 and OQ-05 logged.

+++ services/settlement-service/src/main/java/com/northstar/settlement/domain/SettlementResult.java
@@ -1,8 +1,6 @@
 package com.northstar.settlement.domain;
 
-import java.math.BigDecimal;
-
-public record SettlementResult(BigDecimal coveredAmount,
-        BigDecimal deductibleApplied, BigDecimal depreciation,
-        boolean cappedAtLimit, BigDecimal settlementAmount) {
+/** The legacy Settlement value object minus its identity columns, on double. */
+public record SettlementResult(double coveredAmount, double deductibleApplied,
+        double depreciation, boolean cappedAtLimit, double settlementAmount) {
 }
